Fixes issue #2424 - Add Jacoco check to welcome-file extension
