+++ extension/standard/src/main/java/module-info.java
@@ -45,10 +45,10 @@
  *  <li>Java Policy</li>
  *  <li>Servlet Annotations</li>
  *  <li>Servlet Container Initializer</li>
- *  <li>Welcome File</li>
  *  <li>TEMPDIR</li>
  *  <li>WaSP (Pages)</li>
  *  <li>web.xml</li>
+ *  <li>Welcome File</li>
  * </ul>
  */
 module cloud.piranha.extension.standard {
@@ -66,8 +66,8 @@ module cloud.piranha.extension.standard {
     requires cloud.piranha.extension.standard.policy;
     requires cloud.piranha.extension.standard.servletannotations;
     requires cloud.piranha.extension.standard.scinitializer;
-    requires cloud.piranha.extension.standard.welcomefile;
     requires cloud.piranha.extension.tempdir;
     requires cloud.piranha.extension.wasp;
     requires cloud.piranha.extension.webxml;
+    requires cloud.piranha.extension.welcomefile;
 }
