+++ extension/webxml/src/test/java/cloud/piranha/extension/webxml/WelcomeFileTest.java
@@ -28,7 +28,7 @@
 package cloud.piranha.extension.webxml;
 
 import cloud.piranha.core.impl.DefaultWebApplication;
-import cloud.piranha.extension.standard.welcomefile.StandardWelcomeFileManager;
+import cloud.piranha.extension.welcomefile.WelcomeFileManager;
 import cloud.piranha.resource.impl.DirectoryResource;
 import java.io.File;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -51,7 +51,7 @@ class WelcomeFileTest {
     void testGetWelcomeFileList() throws Exception {
         DefaultWebApplication webApplication = new DefaultWebApplication();
         webApplication.addResource(new DirectoryResource(new File("src/test/webxml/welcomeFile")));
-        webApplication.getManager().setWelcomeFileManager(new StandardWelcomeFileManager());
+        webApplication.getManager().setWelcomeFileManager(new WelcomeFileManager());
         webApplication.addInitializer(new WebXmlInitializer());
         webApplication.initialize();
         assertTrue(webApplication.getManager().getWelcomeFileManager().getWelcomeFileList().contains("index.xhtml"));
@@ -66,7 +66,7 @@ class WelcomeFileTest {
     void testGetWelcomeFileList2() throws Exception {
         DefaultWebApplication webApplication = new DefaultWebApplication();
         webApplication.addResource(new DirectoryResource(new File("src/test/webxml/welcomeFile")));
-        webApplication.getManager().setWelcomeFileManager(new StandardWelcomeFileManager());
+        webApplication.getManager().setWelcomeFileManager(new WelcomeFileManager());
         webApplication.addInitializer(new WebXmlInitializer());
         webApplication.initialize();
         assertFalse(webApplication.getManager().getWelcomeFileManager().getWelcomeFileList().contains("index.jsp"));
