+++ extension/welcomefile/pom.xml
@@ -9,22 +9,12 @@
         <version>22.2.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>piranha-extension-webxml</artifactId>
+    <artifactId>piranha-extension-welcomefile</artifactId>
     <packaging>jar</packaging>
 
-    <name>Piranha - Extension - web.xml</name>
-    <properties>
-        <argLine></argLine>
-    </properties>
+    <name>Piranha - Extension - Welcome File</name>
     <build>
         <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                    <argLine>@{argLine} --add-opens cloud.piranha.extension.webxml/cloud.piranha.extension.webxml.internal=ALL-UNNAMED</argLine>
-                </configuration>
-            </plugin> 
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
@@ -55,7 +45,7 @@
                         </configuration>
                     </execution>
                 </executions>            
-            </plugin>
+            </plugin>            
         </plugins>
     </build>
     <dependencies>
@@ -66,46 +56,11 @@
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
+        <!-- test -->
         <dependency>
             <groupId>cloud.piranha.core</groupId>
             <artifactId>piranha-core-impl</artifactId>
             <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>cloud.piranha.resource</groupId>
-            <artifactId>piranha-resource-impl</artifactId>
-            <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <!-- test -->
-        <dependency>
-            <groupId>cloud.piranha.extension</groupId>
-            <artifactId>piranha-extension-herring</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>cloud.piranha.extension</groupId>
-            <artifactId>piranha-extension-standard-mimetype</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>cloud.piranha.extension</groupId>
-            <artifactId>piranha-extension-slim-security</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>cloud.piranha.extension</groupId>
-            <artifactId>piranha-extension-welcomefile</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.hsqldb</groupId>
-            <artifactId>hsqldb</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
